Added operator for the clover case to D_psi.h/c. This version compiles and has the main ingredients to be used in the inverter.
