+++ operator/D_psi.c
@@ -46,6 +46,7 @@
 #include "update_backward_gauge.h"
 #include "block.h"
 #include "operator/D_psi.h"
+#include "operator/clovertm_operators.h"
 #include "solver/dirac_operator_eigenvectors.h"
 
 #if (defined SSE23 || defined SSE33)
@@ -266,10 +267,7 @@ void local_H(spinor * const rr, spinor * const s, su3 * u, int * _idx) {
 #else
 
 
-//static inline void p0add(spinor * restrict const tmpr , spinor const * restrict const s, 
-//			 su3 const * restrict const u, const _Complex double phase) {
-
-inline void p0add(spinor * restrict const tmpr , spinor const * restrict const s, 
+static inline void p0add(spinor * restrict const tmpr , spinor const * restrict const s, 
 			 su3 const * restrict const u, const _Complex double phase) {
 #ifdef OMP
 #define static
@@ -297,8 +295,7 @@ inline void p0add(spinor * restrict const tmpr , spinor const * restrict const s
 }
 
 
-//static inline void m0add(spinor * restrict const tmpr, spinor const * restrict const s, 
-inline void m0add(spinor * restrict const tmpr, spinor const * restrict const s, 
+static inline void m0add(spinor * restrict const tmpr, spinor const * restrict const s, 
 			 su3 const * restrict const u, const _Complex double phase) {
 #ifdef OMP
 #define static
@@ -325,8 +322,7 @@ inline void m0add(spinor * restrict const tmpr, spinor const * restrict const s,
   return;
 }
 
-//static inline void p1add(spinor * restrict const tmpr, spinor const * restrict const s, 
-inline void p1add(spinor * restrict const tmpr, spinor const * restrict const s, 
+static inline void p1add(spinor * restrict const tmpr, spinor const * restrict const s, 
 			 su3 const * restrict const u, const _Complex double phase) {
 #ifdef OMP
 #define static
@@ -353,8 +349,7 @@ inline void p1add(spinor * restrict const tmpr, spinor const * restrict const s,
   return;
 }
 
-//static inline void m1add(spinor * restrict const tmpr, spinor const * restrict const s, 
-inline void m1add(spinor * restrict const tmpr, spinor const * restrict const s, 
+static inline void m1add(spinor * restrict const tmpr, spinor const * restrict const s, 
 			 su3 const * restrict const u, const _Complex double phase) {
 #ifdef OMP
 #define static
@@ -381,8 +376,7 @@ inline void m1add(spinor * restrict const tmpr, spinor const * restrict const s,
   return;
 }
 
-//static inline void p2add(spinor * restrict const tmpr, spinor const * restrict const s, 
-inline void p2add(spinor * restrict const tmpr, spinor const * restrict const s, 
+static inline void p2add(spinor * restrict const tmpr, spinor const * restrict const s, 
 			 su3 const * restrict const u, const _Complex double phase) {
 #ifdef OMP
 #define static
@@ -410,8 +404,7 @@ inline void p2add(spinor * restrict const tmpr, spinor const * restrict const s,
   return;
 }
 
-//static inline void m2add(spinor * restrict const tmpr, spinor const * restrict const s, 
-inline void m2add(spinor * restrict const tmpr, spinor const * restrict const s, 
+static inline void m2add(spinor * restrict const tmpr, spinor const * restrict const s, 
 			 su3 const * restrict const u, const _Complex double phase) {
 #ifdef OMP
 #define static
@@ -438,8 +431,7 @@ inline void m2add(spinor * restrict const tmpr, spinor const * restrict const s,
   return;
 }
 
-//static inline void p3add(spinor * restrict const tmpr, spinor const * restrict const s, 
-inline void p3add(spinor * restrict const tmpr, spinor const * restrict const s, 
+static inline void p3add(spinor * restrict const tmpr, spinor const * restrict const s, 
 			 su3 const * restrict const u, const _Complex double phase) {
 #ifdef OMP
 #define static
@@ -466,8 +458,7 @@ inline void p3add(spinor * restrict const tmpr, spinor const * restrict const s,
   return;
 }
 
-//static inline void m3addandstore(spinor * restrict const r, spinor const * restrict const s, 
-inline void m3addandstore(spinor * restrict const r, spinor const * restrict const s, 
+static inline void m3addandstore(spinor * restrict const r, spinor const * restrict const s, 
 				 su3 const * restrict const u, const _Complex double phase,
          spinor const * restrict const tmpr) {
 #ifdef OMP
@@ -496,8 +487,7 @@ inline void m3addandstore(spinor * restrict const r, spinor const * restrict con
 }
 
 /* this is the hopping part only */
-//static inline void local_H(spinor * const rr, spinor const * const s, su3 const * restrict u, int * _idx, spinor * const restrict tmpr) {
-inline void local_H(spinor * const rr, spinor const * const s, su3 const * restrict u, int * _idx, spinor * const restrict tmpr) {
+static inline void local_H(spinor * const rr, spinor const * const s, su3 const * restrict u, int * _idx, spinor * const restrict tmpr) {
 
   int * idx = _idx;
 
@@ -544,7 +534,7 @@ inline void local_H(spinor * const rr, spinor const * const s, su3 const * restr
 void D_psi(spinor * const P, spinor * const Q){
 
   if(P==Q){
-    printf("Error in D_psi (operator.c):\n");
+    printf("Error in D_psi (D_psi.c):\n");
     printf("Arguments must be differen spinor fields\n");
     printf("Program aborted\n");
     exit(1);
@@ -944,6 +934,423 @@ void D_psi(spinor * const P, spinor * const Q){
 #endif
 }
 
+void Dsw_psi(spinor * const P, spinor * const Q){
+
+  if(P==Q){
+    printf("Error in Dsw_psi (D_psi.c):\n");
+    printf("Arguments must be differen spinor fields\n");
+    printf("Program aborted\n");
+    exit(1);
+  }
+
+#ifdef _GAUGE_COPY2
+  if(g_update_gauge_copy) {
+      update_backward_gauge(g_gauge_field);
+  }
+#endif
+
+# if defined MPI
+  xchange_lexicfield(Q);
+# endif
+
+#ifdef OMP
+#pragma omp parallel
+  {
+#endif
+  int ix,iy,iz;
+  su3 *up,*um;
+  spinor *s,*sp,*sm,*rn;
+  //_Complex double fact1, fact2;
+  spinor rs __attribute__ ((aligned (16)));
+  spinor ALIGN stmp;
+
+  //fact1 = 1. + g_mu * I;
+  //fact2 = conj(fact1);
+
+#ifndef OMP
+  iy=g_iup[0][0];
+  sp=(spinor *) Q + iy;
+  up=&g_gauge_field[0][0];
+#endif
+
+  /************************ loop over all lattice sites *************************/
+#ifdef OMP
+#pragma omp for
+#endif
+  for (ix=0;ix<VOLUME;ix++){
+#ifdef OMP
+    iy=g_iup[ix][0];
+    up=&g_gauge_field[ix][0];
+    sp=(spinor *) Q + iy;
+#endif
+    s=(spinor *) Q + ix;
+    _prefetch_spinor(s);
+
+    /******************************* direction +0 *********************************/
+
+    iy=g_idn[ix][0];
+      
+    sm = (spinor *) Q + iy;
+    _prefetch_spinor(sm);       
+
+    _sse_load(sp->s0);
+    _sse_load_up(sp->s2);
+    _sse_vector_add();
+
+    _sse_su3_multiply((*up));
+    _sse_vector_cmplx_mul(phase_0);
+    _sse_store_up(rs.s2);
+
+    
+     assign_mul_one_sw_pm_imu_site_lexic(ix,&stmp,s,g_mu);
+
+    _sse_load_up(stmp.s0);
+//    _sse_load_up(s->s0);
+//    _sse_vector_cmplx_mul(fact1);
+/*     _sse_vector_mul(fact1); */
+    _sse_load(rs.s2);
+    _sse_vector_add();
+    _sse_store(rs.s0);
+
+    _sse_load_up(stmp.s2);
+//    _sse_load_up(s->s2);
+//    _sse_vector_cmplx_mul(fact2);
+/*     _sse_vector_mul(fact1);       */
+    _sse_load(rs.s2);
+    _sse_vector_add();
+    _sse_store(rs.s2);      
+      
+    um=&g_gauge_field[iy][0];
+    _prefetch_su3(um);
+      
+    _sse_load(sp->s1);
+    _sse_load_up(sp->s3);
+    _sse_vector_add();
+      
+    _sse_su3_multiply((*up));
+    _sse_vector_cmplx_mul(phase_0);
+    _sse_store_up(rs.s3);
+    
+    _sse_load_up(stmp.s1);
+//    _sse_load_up(s->s1);
+//    _sse_vector_cmplx_mul(fact1);
+/*     _sse_vector_mul(fact1); */
+    _sse_load(rs.s3);
+    _sse_vector_add();
+    _sse_store(rs.s1);
+
+    _sse_load_up(stmp.s3);
+//    _sse_load_up(s->s3);
+//    _sse_vector_cmplx_mul(fact2);
+/*     _sse_vector_mul(fact1);       */
+    _sse_load(rs.s3);
+    _sse_vector_add();
+    _sse_store(rs.s3); 
+
+    /******************************* direction -0 *********************************/
+
+    iy=g_iup[ix][1];
+
+    sp = (spinor *) Q + iy;
+    _prefetch_spinor(sp);
+
+    _sse_load(sm->s0);
+    _sse_load_up(sm->s2);
+    _sse_vector_sub();
+      
+    _sse_su3_inverse_multiply((*um));
+    _sse_vector_cmplxcg_mul(phase_0);
+    _sse_load(rs.s0);
+    _sse_vector_add();
+    _sse_store(rs.s0);
+
+    _sse_load(rs.s2);
+    _sse_vector_sub();
+    _sse_store(rs.s2);
+      
+    up+=1;
+    _prefetch_su3(up);
+      
+    _sse_load(sm->s1);
+    _sse_load_up(sm->s3);
+    _sse_vector_sub();
+      
+    _sse_su3_inverse_multiply((*um));
+    _sse_vector_cmplxcg_mul(phase_0);
+    _sse_load(rs.s1);
+    _sse_vector_add();
+    _sse_store(rs.s1);
+
+    _sse_load(rs.s3);
+    _sse_vector_sub();
+    _sse_store(rs.s3);
+      
+    /******************************* direction +1 *********************************/
+
+    iy=g_idn[ix][1];
+      
+    sm = (spinor *) Q + iy;
+    _prefetch_spinor(sm);
+
+    _sse_load(sp->s0);
+    _sse_load_up(sp->s3);
+    _sse_vector_i_mul();
+    _sse_vector_add();
+
+    _sse_su3_multiply((*up));
+    _sse_vector_cmplx_mul(phase_1);
+    _sse_load(rs.s0);
+    _sse_vector_add();
+    _sse_store(rs.s0);
+
+    _sse_load(rs.s3);
+    _sse_vector_i_mul();      
+    _sse_vector_sub();
+    _sse_store(rs.s3); 
+      
+    um=&g_gauge_field[iy][1];
+    _prefetch_su3(um);
+
+    _sse_load(sp->s1);
+    _sse_load_up(sp->s2);
+    _sse_vector_i_mul();
+    _sse_vector_add();
+
+    _sse_su3_multiply((*up));
+    _sse_vector_cmplx_mul(phase_1);
+    _sse_load(rs.s1);
+    _sse_vector_add();
+    _sse_store(rs.s1);
+
+    _sse_load(rs.s2);
+    _sse_vector_i_mul();      
+    _sse_vector_sub();
+    _sse_store(rs.s2);       
+
+    /******************************* direction -1 *********************************/
+
+    iy=g_iup[ix][2];
+
+    sp = (spinor *) Q + iy;
+    _prefetch_spinor(sp);
+
+    _sse_load(sm->s0);
+    _sse_load_up(sm->s3);
+    _sse_vector_i_mul();
+    _sse_vector_sub();
+      
+    _sse_su3_inverse_multiply((*um));
+    _sse_vector_cmplxcg_mul(phase_1);
+    _sse_load(rs.s0);
+    _sse_vector_add();
+    _sse_store(rs.s0);
+
+    _sse_load(rs.s3);
+    _sse_vector_i_mul();      
+    _sse_vector_add();
+    _sse_store(rs.s3);
+
+    up+=1;
+    _prefetch_su3(up);
+
+    _sse_load(sm->s1);
+    _sse_load_up(sm->s2);
+    _sse_vector_i_mul();
+    _sse_vector_sub();
+      
+    _sse_su3_inverse_multiply((*um));
+    _sse_vector_cmplxcg_mul(phase_1);
+    _sse_load(rs.s1);
+    _sse_vector_add();
+    _sse_store(rs.s1);
+
+    _sse_load(rs.s2);
+    _sse_vector_i_mul();      
+    _sse_vector_add();
+    _sse_store(rs.s2);
+
+    /******************************* direction +2 *********************************/
+
+    iy=g_idn[ix][2];
+
+    sm = (spinor *) Q + iy;
+    _prefetch_spinor(sm);
+
+    _sse_load(sp->s0);
+    _sse_load_up(sp->s3);
+    _sse_vector_add();
+
+    _sse_su3_multiply((*up));
+    _sse_vector_cmplx_mul(phase_2);
+    _sse_load(rs.s0);
+    _sse_vector_add();
+    _sse_store(rs.s0);
+
+    _sse_load(rs.s3);
+    _sse_vector_add();
+    _sse_store(rs.s3);
+      
+    um=&g_gauge_field[iy][2];
+    _prefetch_su3(um);
+
+    _sse_load(sp->s1);
+    _sse_load_up(sp->s2);
+    _sse_vector_sub();
+
+    _sse_su3_multiply((*up));
+    _sse_vector_cmplx_mul(phase_2);
+    _sse_load(rs.s1);
+    _sse_vector_add();
+    _sse_store(rs.s1);
+
+    _sse_load(rs.s2);
+    _sse_vector_sub();
+    _sse_store(rs.s2);      
+
+    /******************************* direction -2 *********************************/
+
+    iy=g_iup[ix][3];
+
+    sp = (spinor *) Q + iy;
+    _prefetch_spinor(sp);
+
+    _sse_load(sm->s0);
+    _sse_load_up(sm->s3);
+    _sse_vector_sub();
+      
+    _sse_su3_inverse_multiply((*um));
+    _sse_vector_cmplxcg_mul(phase_2);
+    _sse_load(rs.s0);
+    _sse_vector_add();
+    _sse_store(rs.s0);
+
+    _sse_load(rs.s3);
+    _sse_vector_sub();
+    _sse_store(rs.s3);
+      
+    up+=1;
+    _prefetch_su3(up);
+
+    _sse_load(sm->s1);
+    _sse_load_up(sm->s2);
+    _sse_vector_add();
+      
+    _sse_su3_inverse_multiply((*um));
+    _sse_vector_cmplxcg_mul(phase_2);
+    _sse_load(rs.s1);
+    _sse_vector_add();
+    _sse_store(rs.s1);
+
+    _sse_load(rs.s2);
+    _sse_vector_add();
+    _sse_store(rs.s2);      
+      
+    /******************************* direction +3 *********************************/
+
+    iy=g_idn[ix][3];
+
+    sm = (spinor *) Q + iy;
+    _prefetch_spinor(sm);
+
+    _sse_load(sp->s0);
+    _sse_load_up(sp->s2);
+    _sse_vector_i_mul();
+    _sse_vector_add();
+
+    _sse_su3_multiply((*up));
+    _sse_vector_cmplx_mul(phase_3);
+    _sse_load(rs.s0);
+    _sse_vector_add();
+    _sse_store(rs.s0);
+
+    _sse_load(rs.s2);
+    _sse_vector_i_mul();      
+    _sse_vector_sub();
+    _sse_store(rs.s2);
+      
+    um=&g_gauge_field[iy][3];
+    _prefetch_su3(um);
+
+    _sse_load(sp->s1);
+    _sse_load_up(sp->s3);
+    _sse_vector_i_mul();
+    _sse_vector_sub();
+
+    _sse_su3_multiply((*up));
+    _sse_vector_cmplx_mul(phase_3);
+    _sse_load(rs.s1);
+    _sse_vector_add();
+    _sse_store(rs.s1);
+
+    _sse_load(rs.s3);
+    _sse_vector_i_mul();      
+    _sse_vector_add();
+    _sse_store(rs.s3);
+      
+    /******************************* direction -3 *********************************/
+
+    iz=(ix+1+VOLUME)%VOLUME;
+
+    iy=g_iup[iz][0];
+      
+    sp = (spinor *) Q + iy;
+    _prefetch_spinor(sp);
+
+    _sse_load(sm->s0);
+    _sse_load_up(sm->s2);
+    _sse_vector_i_mul();
+    _sse_vector_sub();
+      
+    _sse_su3_inverse_multiply((*um));
+    _sse_vector_cmplxcg_mul(phase_3);
+    rn = (spinor *) P + ix;
+      
+    _sse_load(rs.s0);
+    _sse_vector_add();
+/*     _sse_vector_mul(fact2); */
+    _sse_store_nt(rn->s0);
+
+    _sse_load(rs.s2);
+    _sse_vector_i_mul();      
+    _sse_vector_add();
+/*     _sse_vector_mul(fact2);       */
+    _sse_store_nt(rn->s2);
+
+    up=&g_gauge_field[iz][0];
+    _prefetch_su3(up);
+
+    _sse_load(sm->s1);
+    _sse_load_up(sm->s3);
+    _sse_vector_i_mul();
+    _sse_vector_add();
+      
+    _sse_su3_inverse_multiply((*um));
+    _sse_vector_cmplxcg_mul(phase_3);
+    _sse_load(rs.s1);
+    _sse_vector_add();
+/*     _sse_vector_mul(fact2);       */
+    _sse_store_nt(rn->s1);
+
+    _sse_load(rs.s3);
+    _sse_vector_i_mul();      
+    _sse_vector_sub();
+/*     _sse_vector_mul(fact2); */
+    _sse_store_nt(rn->s3);
+      
+    /******************************** end of loop *********************************/
+
+  }
+#ifdef OMP
+  } /* OpenMP closing brace */
+#endif
+}
+
+
+
+
+
+
+
 #elif ((defined BGL) && (defined XLC))
 
 
@@ -1202,6 +1609,259 @@ void D_psi(spinor * const P, spinor * const Q){
 }
 
 
+void Dsw_psi(spinor * const P, spinor * const Q){
+  int ix,iy,iz;
+  //static _Complex double fact1;
+  su3 * restrict up ALIGN;
+  su3 * restrict um ALIGN;
+  spinor * restrict s ALIGN;
+  spinor * restrict sp ALIGN;
+  spinor * restrict sm ALIGN;
+  spinor * restrict rn ALIGN;
+  spinor ALIGN stmp;
+
+#pragma disjoint(*s, *sp, *sm, *rn, *up, *um, *P, *Q)
+
+  __alignx(16,P);
+  __alignx(16,Q);
+
+#ifdef _GAUGE_COPY
+  if(g_update_gauge_copy) {
+    update_backward_gauge(g_gauge_field);
+  }
+#endif
+
+#    if (defined MPI && !(defined _NO_COMM))
+  xchange_lexicfield(Q);
+#    endif
+
+  //fact1 = 1.0 + g_mu * I;
+
+  iy=g_iup[0][0];
+  sp=(spinor *) Q + iy;
+  up=&g_gauge_field[0][0];
+
+  /**************** loop over all lattice sites ******************/
+  for(ix = 0; ix < VOLUME; ix++){
+    s=(spinor *) Q + ix;
+    rn = (spinor *) P + ix;
+    assign_mul_one_sw_pm_imu_site_lexic(ix,&stmp,s,g_mu);
+    /*********************** direction +0 ************************/
+
+    iy=g_idn[ix][0]; 
+
+    um=&g_gauge_field[iy][0]; 
+
+    _prefetch_su3(um); 
+    sm = (spinor*) Q + iy;
+    _prefetch_spinor(sm); 
+
+    _bgl_load_reg0(sp->s0);
+    _bgl_load_reg1(sp->s1);
+    _bgl_load_reg0_up(sp->s2);
+    _bgl_load_reg1_up(sp->s3);
+    _bgl_vector_add_reg0();
+    _bgl_vector_add_reg1();
+    /* result is now in regx0, regx1, regx2 x = 0,1 */
+
+    _bgl_su3_multiply_double((*up));
+    _bgl_vector_cmplx_mul_double(phase_0);
+    //_bgl_load_rs0(s->s0);
+    //_bgl_load_rs1(s->s1);
+    //_bgl_load_rs2(s->s2);
+    //_bgl_load_rs3(s->s3);
+    //_bgl_vector_cmplx_mul_rs(fact1);
+    _bgl_load_rs0(stmp.s0);
+    _bgl_load_rs1(stmp.s1);
+    _bgl_load_rs2(stmp.s2);
+    _bgl_load_rs3(stmp.s3);
+    
+    _bgl_add_to_rs0_reg0();
+    _bgl_add_to_rs2_reg0();
+    _bgl_add_to_rs1_reg1();
+    _bgl_add_to_rs3_reg1();
+
+    /*********************** direction -0 ************************/
+
+    iy=g_iup[ix][1]; 
+
+    up+=1;
+    _prefetch_su3(up); 
+    sp = (spinor *) Q + iy;
+    _prefetch_spinor(sp); 
+
+    _bgl_load_reg0(sm->s0);
+    _bgl_load_reg1(sm->s1);
+    _bgl_load_reg0_up(sm->s2);
+    _bgl_load_reg1_up(sm->s3);
+    _bgl_vector_sub_reg0();
+    _bgl_vector_sub_reg1();
+
+    _bgl_su3_inverse_multiply_double((*um));
+    _bgl_vector_cmplxcg_mul_double(phase_0);
+
+    _bgl_add_to_rs0_reg0();
+    _bgl_sub_from_rs2_reg0();
+    _bgl_add_to_rs1_reg1();
+    _bgl_sub_from_rs3_reg1();
+
+    /*********************** direction +1 ************************/
+
+    iy=g_idn[ix][1]; 
+
+    um=&g_gauge_field[iy][1]; 
+
+    _prefetch_su3(um); 
+    sm = (spinor *) Q + iy;
+    _prefetch_spinor(sm); 
+
+    _bgl_load_reg0(sp->s0);
+    _bgl_load_reg1(sp->s1);
+    _bgl_load_reg0_up(sp->s3);
+    _bgl_load_reg1_up(sp->s2);
+    _bgl_vector_i_mul_add_reg0();
+    _bgl_vector_i_mul_add_reg1();
+
+    _bgl_su3_multiply_double((*up));
+    _bgl_vector_cmplx_mul_double(phase_1);
+
+    _bgl_add_to_rs0_reg0();
+    _bgl_i_mul_sub_from_rs3_reg0();
+    _bgl_add_to_rs1_reg1();
+    _bgl_i_mul_sub_from_rs2_reg1();
+
+    /*********************** direction -1 ************************/
+
+    iy=g_iup[ix][2]; 
+
+    up+=1;
+    _prefetch_su3(up); 
+    sp = (spinor *) Q + iy;
+    _prefetch_spinor(sp); 
+
+    _bgl_load_reg0(sm->s0);
+    _bgl_load_reg1(sm->s1);
+    _bgl_load_reg0_up(sm->s3);
+    _bgl_load_reg1_up(sm->s2);
+    _bgl_vector_i_mul_sub_reg0();
+    _bgl_vector_i_mul_sub_reg1();
+
+    _bgl_su3_inverse_multiply_double((*um));
+    _bgl_vector_cmplxcg_mul_double(phase_1);
+
+    _bgl_add_to_rs0_reg0();
+    _bgl_add_to_rs1_reg1();
+    _bgl_i_mul_add_to_rs3_reg0();
+    _bgl_i_mul_add_to_rs2_reg1();
+
+    /*********************** direction +2 ************************/
+
+    iy=g_idn[ix][2];
+
+    um=&g_gauge_field[iy][2];
+    _prefetch_su3(um);
+    sm = (spinor *) Q + iy;
+    _prefetch_spinor(sm);
+
+    _bgl_load_reg0(sp->s0);
+    _bgl_load_reg1(sp->s1);
+    _bgl_load_reg1_up(sp->s2);
+    _bgl_load_reg0_up(sp->s3);
+    _bgl_vector_add_reg0();
+    _bgl_vector_sub_reg1();
+
+    _bgl_su3_multiply_double((*up));
+    _bgl_vector_cmplx_mul_double(phase_2);
+
+    _bgl_add_to_rs0_reg0();
+    _bgl_add_to_rs1_reg1();
+    _bgl_sub_from_rs2_reg1();
+    _bgl_add_to_rs3_reg0();
+
+
+    /*********************** direction -2 ************************/
+
+    iy=g_iup[ix][3]; 
+
+    up+=1;
+    _prefetch_su3(up); 
+    sp = (spinor *) Q + iy;
+    _prefetch_spinor(sp); 
+
+    _bgl_load_reg0(sm->s0);
+    _bgl_load_reg1(sm->s1);
+    _bgl_load_reg1_up(sm->s2);
+    _bgl_load_reg0_up(sm->s3);
+    _bgl_vector_sub_reg0();
+    _bgl_vector_add_reg1();
+
+    _bgl_su3_inverse_multiply_double((*um));
+    _bgl_vector_cmplxcg_mul_double(phase_2);
+
+    _bgl_add_to_rs0_reg0();
+    _bgl_add_to_rs1_reg1();
+    _bgl_add_to_rs2_reg1();
+    _bgl_sub_from_rs3_reg0();
+
+    /*********************** direction +3 ************************/
+
+    iy=g_idn[ix][3]; 
+
+    um=&g_gauge_field[iy][3]; 
+    _prefetch_su3(um); 
+    sm = (spinor *) Q + iy;
+    _prefetch_spinor(sm); 
+
+    _bgl_load_reg0(sp->s0);
+    _bgl_load_reg1(sp->s1);
+    _bgl_load_reg0_up(sp->s2);
+    _bgl_load_reg1_up(sp->s3);
+    _bgl_vector_i_mul_add_reg0();
+    _bgl_vector_i_mul_sub_reg1();
+
+    _bgl_su3_multiply_double((*up));
+    _bgl_vector_cmplx_mul_double(phase_3);
+
+    _bgl_add_to_rs0_reg0();
+    _bgl_add_to_rs1_reg1();
+    _bgl_i_mul_sub_from_rs2_reg0();
+    _bgl_i_mul_add_to_rs3_reg1();
+
+    /*********************** direction -3 ************************/
+
+    iz=(ix+1+VOLUME)%VOLUME;
+
+    iy=g_iup[iz][0];
+
+    up=&g_gauge_field[iz][0];
+    _prefetch_su3(up); 
+    sp = (spinor *) Q + iy;
+    _prefetch_spinor(sp); 
+
+    _bgl_load_reg0(sm->s0);
+    _bgl_load_reg1(sm->s1);
+    _bgl_load_reg0_up(sm->s2);
+    _bgl_load_reg1_up(sm->s3);
+    _bgl_vector_i_mul_sub_reg0();
+    _bgl_vector_i_mul_add_reg1();
+
+    _bgl_su3_inverse_multiply_double((*um));
+    _bgl_vector_cmplxcg_mul_double(phase_3);
+
+    _bgl_add_to_rs0_reg0();
+    _bgl_store_rs0(rn->s0);
+    _bgl_i_mul_add_to_rs2_reg0();
+    _bgl_store_rs2(rn->s2);
+
+    _bgl_add_to_rs1_reg1();
+    _bgl_store_rs1(rn->s1);
+    _bgl_i_mul_sub_from_rs3_reg1();
+    _bgl_store_rs3(rn->s3);
+
+    /************************ end of loop ************************/
+  }
+}
+
 #else
 
 /* Serially Checked ! */
@@ -1253,6 +1913,107 @@ void D_psi(spinor * const P, spinor * const Q){
     _complex_times_vector(tmpr.s1, rho1, s->s1);
     _complex_times_vector(tmpr.s2, rho2, s->s2);
     _complex_times_vector(tmpr.s3, rho2, s->s3);
+
+    /******************************* direction +0 *********************************/
+    iy=g_iup[ix][0];
+    sp = (spinor *) Q +iy;
+    up=&g_gauge_field[ix][0];
+    p0add(&tmpr, sp, up, phase_0);
+
+    /******************************* direction -0 *********************************/
+    iy=g_idn[ix][0];
+    sm  = (spinor *) Q +iy;
+    um=&g_gauge_field[iy][0];
+    m0add(&tmpr, sm, um, phase_0);
+
+    /******************************* direction +1 *********************************/
+    iy=g_iup[ix][1];
+    sp = (spinor *) Q +iy;
+    up=&g_gauge_field[ix][1];
+    p1add(&tmpr, sp, up, phase_1);
+
+    /******************************* direction -1 *********************************/
+    iy=g_idn[ix][1];
+    sm = (spinor *) Q +iy;
+    um=&g_gauge_field[iy][1];
+    m1add(&tmpr, sm, um, phase_1);
+
+    /******************************* direction +2 *********************************/
+    iy=g_iup[ix][2];
+    sp = (spinor *) Q +iy;
+    up=&g_gauge_field[ix][2];
+    p2add(&tmpr, sp, up, phase_2);
+
+    /******************************* direction -2 *********************************/
+    iy=g_idn[ix][2];
+    sm = (spinor *) Q +iy;
+    um=&g_gauge_field[iy][2];
+    m2add(&tmpr, sm, um, phase_2);
+
+    /******************************* direction +3 *********************************/
+    iy=g_iup[ix][3];
+    sp = (spinor *) Q +iy;
+    up=&g_gauge_field[ix][3];
+    p3add(&tmpr, sp, up, phase_3);
+
+    /******************************* direction -3 *********************************/
+    iy=g_idn[ix][3];
+    sm = (spinor *) Q +iy;
+    um=&g_gauge_field[iy][3];
+    m3addandstore(rr, sm, um, phase_3, &tmpr);
+  }
+#ifdef OMP
+  } /* OpenMP closing brace */
+#endif
+}
+
+void Dsw_psi(spinor * const P, spinor * const Q){
+  if(P==Q){
+    printf("Error in Dsw_psi (Dsw_psi.c):\n");
+    printf("Arguments must be different spinor fields\n");
+    printf("Program aborted\n");
+    exit(1);
+  }
+#ifdef _GAUGE_COPY
+  if(g_update_gauge_copy) {
+      update_backward_gauge(g_gauge_field);
+  }
+#endif
+# if defined MPI
+  xchange_lexicfield(Q);
+# endif
+
+#ifdef OMP
+#pragma omp parallel
+  {
+#endif
+
+  int ix,iy;
+  su3 * restrict up,* restrict um;
+  spinor * restrict rr; 
+  spinor const * restrict s;
+  spinor const * restrict sp;
+  spinor const * restrict sm;
+  //_Complex double rho1, rho2;
+  spinor ALIGN tmpr;
+
+  //rho1 = 1. + g_mu * I;
+  //rho2 = conj(rho1);
+
+  /************************ loop over all lattice sites *************************/
+
+#ifdef OMP
+#pragma omp for
+#endif
+  for (ix=0;ix<VOLUME;ix++)
+  {
+    rr  = (spinor *) P +ix;
+    s  = (spinor *) Q +ix;
+    assign_mul_one_sw_pm_imu_site_lexic(ix,&tmpr,s,g_mu);
+    //_complex_times_vector(tmpr.s0, rho1, s->s0);
+    //_complex_times_vector(tmpr.s1, rho1, s->s1);
+    //_complex_times_vector(tmpr.s2, rho2, s->s2);
+    //_complex_times_vector(tmpr.s3, rho2, s->s3);
 
     /******************************* direction +0 *********************************/
     iy=g_iup[ix][0];
@@ -1373,6 +2134,107 @@ void Block_D_psi(block * blk, spinor * const rr, spinor * const s) {
   }
   return;
 }
+
+
+void Dsw_psi_prec(spinor * const P, spinor * const Q){
+
+  /* todo: do preconditioning */
+  spinorPrecWS *ws=(spinorPrecWS*)g_precWS;
+  static _Complex double alpha = -1.0;
+
+  alpha = -0.5;
+  spinorPrecondition(P,Q,ws,T,L,alpha,0,1);
+  Dsw_psi(g_spinor_field[DUM_MATRIX],P);
+  alpha = -0.5;
+  spinorPrecondition(P,g_spinor_field[DUM_MATRIX],ws,T,L,alpha,0,1);
+}
+
+/* apply the Dirac operator to the block local spinor field s */
+/* and store the result in block local spinor field rr        */
+/* for block blk                                              */
+/* the block local gauge field is assumed to be in the order  */
+/* that is needed int local_D, which means also that it is a  */
+/* double copy                                                */
+// CU: has problems with SSE2,3
+void Block_Dsw_psi(block * blk, spinor * const rr, spinor * const s) {
+  int i;
+  spinor *r = rr;
+  spinor *t = s;
+  su3 * u = blk->u;
+  int * idx = blk->idx;
+  //static _Complex double rhoa, rhob;
+  spinor tmpr;
+
+  int it,ix,iy,iz; //lexiographic index of the site w.r.t the block
+  int bt,bx,by,bz; //block coordinate on the local mpi process
+  int dT,dX,dY,dZ; //block size
+  int sT,sX,sY,sZ; //constant shifts
+  int lx; //lexiographic index of the block site w.r.t the local mpi process
+
+  dT = blk->BT;
+  dX = blk->BLX;
+  dY = blk->BLY;
+  dZ = blk->BLZ;
+
+  bt = blk->mpilocal_coordinate[0];
+  bx = blk->mpilocal_coordinate[1];
+  by = blk->mpilocal_coordinate[2];
+  bz = blk->mpilocal_coordinate[3];
+
+  sT = bt*dT;
+  sX = bx*dX;
+  sY = by*dY;
+  sZ = bz*dZ;
+  
+#if (defined BGL && defined XLC)
+  __alignx(16,s);
+#endif
+  if(blk_gauge_eo) {
+    init_blocks_gaugefield();
+  }
+  //rhoa = 1.0 + g_mu * I;
+  //rhob = conj(rhoa);
+
+  /* set the boundary term to zero */
+  _spinor_null(rr[blk->volume]);
+  _spinor_null(s[blk->volume]);
+
+  for(i = 0; i < blk->volume; i++) {
+
+     iz = i%dZ;
+     iy = (i/dZ)%dY;
+     ix = (i/(dZ*dY))%dX;
+     it = i/(dZ*dY*dX);
+
+     lx = g_ipt[it+sT][ix+sX][iy+sY][iz+sZ];
+
+     assign_mul_one_sw_pm_imu_site_lexic(lx,&tmpr,t,g_mu);
+
+/*
+#if (defined BGL && defined XLC)
+    _bgl_load_rs0(t->s0);
+    _bgl_load_rs1(t->s1);
+    _bgl_load_rs2(t->s2);
+    _bgl_load_rs3(t->s3);
+    _bgl_vector_cmplx_mul_rs(rhoa);
+#else
+    _complex_times_vector(tmpr.s0, rhoa, t->s0);
+    _complex_times_vector(tmpr.s1, rhoa, t->s1);
+    _complex_times_vector(tmpr.s2, rhob, t->s2);
+    _complex_times_vector(tmpr.s3, rhob, t->s3);
+#endif
+*/
+    local_H(r, s, u, idx, &tmpr);
+
+    r++;
+    t++;
+    idx += 8;
+    u += 8;
+  }
+
+  return;
+}
+
 
 /* Apply Hopping Matrix to a even(odd) spinor */
 void Block_H_psi(block * blk, spinor * const rr, spinor * const s, const int eo) {
