--- operator/Dsw_psi.h
@@ -1,29 +1,0 @@
-/***********************************************************************
- *
- * Copyright (C) 2007,2008 Carsten Urbach
- *
- * This file is part of tmLQCD.
- *
- * tmLQCD is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * tmLQCD is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with tmLQCD.  If not, see <http://www.gnu.org/licenses/>.
- ***********************************************************************/
-
-#ifndef _DSW_PSI_H
-#define _DSW_PSI_H
-
-#include "block.h"
-
-void Dsw_psi(spinor * const P, spinor * const Q);
-void Dsw_psi_prec(spinor * const P, spinor * const Q);
-
-#endif
